[#126][refactoring] Extract production code
